Fix some intermittent test failures
Some tests leak Skylight::Instrumenter.current_trace, this can cause
future tests to fail. This commit attempts to clean up these cases, but
the problem should be revisited.

Since trace builders hold a reference to the instrumenter, it should be
possible to check whether or not the instrumenter has been dropped and
restarted and lazily clear stale trace builder objects.